Fix `restoreConversation` function to save only user-specific tasks
Previously, the `restoreConversation` function contained a nested
`map` function that returned an array of Task objects each of which
contained an array of Conversation objects. However, this returned
array contained all the Task objects in the Botkit instance instead
of only the Task objects related to a specific user id.

Therefore, the array traversal had to filter out Tasks and Conversations
not related to a given user id. Implementing the `map` function together
with the filtering would have resulted in an array containing `null`
elements should the array have contained Task objects not related to the
given user id, which is why a `reduce` function was implemented instead.

Signed-off-by: Kenneth Gitere <gitere81@gmail.com>