Add unique identifier for Task objects. Fix having excess instances of Task objects during retrieval from Redis.
This identifier is used to relate Task objects saved on Redis
to the Task objects stored in the tasks array of the Botkit
instance. The identifier itself is a `Symbol` object which
is always unique with every call to it.

By using the identifier, during retrieval of Conversations and
Tasks, formerly the recovered Task would be added into
the array of existing Tasks meaning for every restoration, there
is a new Task object that will be added to the `botkit.tasks`
array. Now the recovered Task only needs to assign its value
to its relative Task object in `botkit.tasks`.
If no tasks are found, for example when restarting the bot,
the recovered Task will then just add itself to `botkit.tasks`.

Signed-off-by: Kenneth Gitere <gitere81@gmail.com>