Add `deleteTask` and `deleteConversation`
The `deleteTask` currently removes tasks that have ended with a
status of `completed`. The conversation is removed from the convos
array of its parent Task object and these changes are persisted to
Redis. In the event of having no more conversations in the convos
array, an empty array is persisted to Redis.

The `deleteConversation` function removes the Task object from the
`botkit.tasks` array and then deletes the field entry of the user
from Redis.

Signed-off-by: Kenneth Gitere <gitere81@gmail.com>