Change task idenifier value from type `Symbol` to type `String`
The primitive Symbol type cannot be directly type cast to a String
hence could not be stored to Redis. When using the `.toString()`
method, for example on the Symbol, Symbol(), the method returns
'Symbol()'. When comparing two stringified Symbols, `'Symbol()'`
is equal to `'Symbol()'`, whereas Symbol() is not equal to Symbol().

Therefore, the identifier is now implemented as a String consisting
of the sum of the user id and the current time in milliseconds.

Signed-off-by: Kenneth Gitere <gitere81@gmail.com>