Fix `storeConversation` when storing the messages array in the Conversation object.
Each message Object literal(Dictionary in Python), contains three
keys: text, channel and handler. The keys: text and channel are of
type String and Number(Integer in other languages) hence can be
type cast to the String type. However, the handler key holds a
value of an anonymous function, particularly a fat arrow function,
which cannot be directly type cast to the String type using the
JSON.stringify() method.

Due to this, JSON.stringify() would omit the handler value when
type casting the message Object literal and instead return undefined.
A catch method was also added after the `redis.hset()` method for any
TypeError errors JSON.stringify() throws.

The absence of the handler key:value pair in an active Conversation
object causes the bot to continue sending out the remaining messages
in the convo.messages array. This results in a premature ending of
a conversation. This occurs in the body of `convo.tick()`

Solution:
The handler value which is a function has a `.toString()` method
returns a string of the entire function hence was used to iterate
over each message Object literal to convert the handler value to
a string. When restoring the Conversation object, the `eval()`
method is called to convert the string to a Function literal.

Signed-off-by: Kenneth Gitere <gitere81@gmail.com>