Persist `convo.sent` and `convo.handler`.
Explanation

When a Conversation instance is created, the value of handler is set
to null and is set to a function by `convo.tick()` when a conversation
is active.
When a message is received, its relative Conversation is searched for
which calls `convo.handle()` to handle the conversation.
Once `convo.handle()` is called, it sets the value of handler to
`null` and the `.tick()` method sets the value of handler to the value
of `message.handler`.

However, this behaviour is disrupted by `storeConversation`, which is
called after a message has been handled but before `convo.tick()` runs.
Therefore, once the message has been handled, the value of handler is
null and is persisted to Redis as null. The array of messages is also
mutated when `convo.tick()` is called but when `storeConversation` is
run, it contains an old version of the array.
The Conversation instance that will be restored afterwards will have
the value of handler set to null which prevents `storeConversation`
from being called, meaning the Conversation instances cannot be updated.

Solution

By default, the newly set value of `convo.handler` is equal to the
value of the handler of the most recent message hence it was explicitly
set when running `storeConversation`.
By setting `convo.handler` to a function, `convo.handle` will call
`convo.capture()` which uses the `convo.sent` array. Therefore, `convo.sent`
also had to be persisted.

Signed-off-by: Kenneth Gitere <gitere81@gmail.com>